[PATCH] console cleanup (2/2)
	Console drivers cleanup.  In current tree interaction between
console and tty layer sits in the ->device() method of struct console.
It takes a pointer to console and returns device number of its tty
device.  open(2) on /dev/console goes through the list of registered
consoles, picks the first one that has ->device() and remaps the device
number to console->device(console).  Then it proceeds with normal
opening of tty.  This is the only caller of ->device().

	Cleanup: let ->device() return a pair (pointer to tty_driver, index
of tty in question) instead of device number.  Note that
	a) the first thing tty_open() does with remapped device number is
conversion to such pair.
	b) console driver _knows_ which tty_driver we want - one that
implements tty interface to the same physical device (i.e. the part of
the same driver).
	c) current code expects the result of ->device() to be a device
number of tty device - anything else is immediate -ENODEV from tty_open();
might as well have NULL ->device in that driver.

	Console drivers converted, (the only) caller updated.

BKrev: 3ea8c9df5d9LOxyVREvWRxPHHBwRDg