Merge flint.arm.linux.org.uk:/usr/src/linux-bk-2.5/linux-2.5-nwfpe into flint.arm.linux.org.uk:/usr/src/linux-bk-2.5/linux-2.5-rmk
2003/04/28 10:19:30+01:00 rmk
[ARM] Clean up nwfpe makefile.

2003/03/29 21:43:13+00:00 rmk
[NWFPE] Clean up indentation in assembly files.

2003/03/29 20:45:38+00:00 ralphs
[NWFPE] performance improvements [Part 6]

NWFPE performance improvements, part 6.  Builds on patch 1468/1.

Break NWFPE initialization into separate function and call it from
entry.S as appropriate.  Also avoid nRc variable usage in EmulateAll().
This buys 6% speedup in NWFPE.  (Note that this is over half of the
total speedup between original RMK and the netwinder cvs version).

(note that 2.5 already has the separate initialisation, and this
cset actually makes 2.5 conform to 2.4 function naming for this
part. --rmk)

2003/03/29 20:14:22+00:00 ralphs
[NWFPE] performance improvements [Part 5]

NWFPE performance improvements, part 5.  Builds on patch 1467/1.

Inline the functions getTransferLength(), getRegisterCount(),
and getRoundingPrecision().  This buys 1.6% speed gain.

2003/03/29 20:03:25+00:00 ralphs
[NWFPE] performance improvements [Part 4]

Fourth part of NWFPE performance improvements, building on patch 1466/1.

This patch makes 80-bit (extended precision) a compile-time option in NWFPE.
The effect of the extra code path on 32/64 bit operations is not as significant
as I had first thought - speedup is between 0.7 and 1.5% depending on which
tree (rmk or nw) you do the changes in.  However the reduction in code size,
particularly when softfloat has the same functions cut out, may be of value
on embedded systems.

2003/03/29 19:46:49+00:00 ralphs
[PATCH] performance improvements [Part 3]

NWFPE performance improvements, part three.

A minor simplification - as (opcode & MASK_ARITHMETIC_OPCODE) >> 20
is frequently used, compute that quantity once and store it.

2003/03/29 18:52:51+00:00 ralphs
[NWFPE] performance improvements [Part 2]

Second part of NWFPE improvements.  Builds on top of patch 1464/1.

Remove unused code.  And fix a function prototype to match function.
No impact on generated code or NWFPE performance.

2003/03/29 16:39:59+00:00 ralphs
[NWFPE] Performance improvements [Part 1]

This is the first of a series of patches to NWFPE, which aim to bring
the performance improvements from the netwinder.org CVS tree into the
mainline ARM kernel.

This patch is merely a reformatting of the NWFPE sources.  All files
were processed with "indent -kr -i8 -ts8 -sob -l132 -ss" and a few
manual fixups. Exception: the softfloat files have not been touched.

This patch should be applied after the 5 patches previously published
by RMK on linux-arm-kernel have been applied.

2003/03/06 15:46:12+00:00 rmk
[ARM] NWFPE 5: Eliminate use of Fd

Slightly better performance can be obtained by eliminating Fd and
calculating it after the math operation.

2003/03/06 15:39:57+00:00 rmk
[ARM] NWFPE 4: Eliminate getFd from CPDO worker functions.

Eliminate getFd from {Single,Double,Extended}CPDO by passing a
pointer to the destination register into these functions.  Use this
same pointer when converting the destination register to the required
type.

2003/03/06 15:36:22+00:00 rmk
[ARM] NWFPE 3: Eliminate setting of fType in CPDO worker functions.

Extending the idea in patch 2, we can eliminate the setting of fType
in {Single,Double,Extended}CPDO functions and always set it in
fpa11_cpdo.c to the ultimate destination size.

2003/03/06 15:31:21+00:00 rmk
[ARM] NWFPE 2: Take advantage of the CPDO functions behaviour.

SingleCPDO, DoubleCPDO and ExtendedCPDO always set the destination
register type to their operating type (single, double, extended
respectively) when they complete successfully (nRc != 0).  This
means that we know what size result will be generated, and we don't
have to fetch it from the fType array.

In fact, since we know "nType" will be the type of the operand,
we don't even need to calculate value of Fd unless we need to
convert the result to the destination type.

2003/03/06 15:26:12+00:00 rmk
[ARM] NWFPE 1: Convert instruction decoding from switch() to table.

BKrev: 3ead8db7xiO2MNUEnYS2KPIAERpmFA