[PATCH] allow modular JBD
From: Paul Clements <Paul.Clements@SteelEye.com>

Currently, when I build ext3 as a module, jbd gets built into the kernel
proper.  This trivial patch allows jbd to be built as a module when ext3 is
also modular.  I believe this is the intention, as this is how it works in
2.4.  I've built and tested with modular jbd and ext3 on 2.5.68.

BKrev: 3eafde6euSIfcnRswDlHXWcbxkK2DA