[PATCH] config menu cleanups
From Robert Day, through "Randy.Dunlap" <rddunlap@osdl.org>

This is a patch from Robert Day that does the following:

1) shift menu item in "Processor type and features" menu
2) clean up "Bus options" menu so it's actually hierarchical

Part of it (moving X86_IO_APIC around) looked a little odd to me,
so I asked Roman Zippel about it, and he replied:

"It's correct, although I wouldn't call it a 'design quirk'. :)
It forces one to group options which belong logically together and in this
case X86_IO_APIC is really a bit misplaced, even if it's not visible."

I have tested it (on 2.5.68-plain) and it does indeed make the menus
more hierarchical.

BKrev: 3eafded2POcP9aVStHtAAPmdyFQD3A