[PATCH] Fix prefetch patching in 2.5-bk
Brown paperbag time. I forgot to take the modrm byte in account
with the prefetch patch replacement.  With 3.2 it worked because
it used the right registers in my configuration.

But gcc 2.96 uses a different register in __dpath and the prefetch becomes
4 bytes with modrm and the original nop needs to be as long as that too.

BKrev: 3eb074ce7Mk6yc3nvkmup19UUW3J5g