[PATCH] hold i_sem on swapfiles
If a swapfile is ftruncated while in use, subsequent swapout will scribble on
the filesystem.

This is a case of root-shoot-foot, but wrecking the fs is a fairly rude
response.  And it's easy to fix: hold i_sem across the life of the swapon.

BKrev: 3eb9e8cf49Jyy13A5v-wHLig0AGBEA