[PATCH] enable slab debugging for larger objects
Some of the fancier slab debugging options are disabled for caches whose
objects are larger than PAGE_SIZE/8, for speed/fragmentation reasons.

But this patch turns up bugs in the size-2048 slab, so it was a bad idea.
Enable the debugging for slabs up to 4096 bytes.

BKrev: 3ec08c22-h5TgYJD2vfgvnWGL-nksw