[PATCH] Framebuffer console fix
This fixes a oops that happens when we map a framebuffer device to a
non-existant console.

set_con2fb_map wasn't testing to see the VC we where mapping to actually
exist.  Now it does.

I also added code to fbcon_cursor to reset the hotspot if it was changed
by userland.

BKrev: 3ec1b04ab_ggwdaVgkGL_xQc6EGgsA