[PATCH] Remove 'strchr' warning from reiserfs
Reiserfs emits a warning about strchr being defined but not used.  I
finally tracked down the reason for this.  gcc - when seeing strstr(x,
"%") recognized that the second parameter is a char, and therefore uses
strchr instead of strstr.  The workaround to avoid the warning is to
replace the call to strstr with strchr - which is OK.

This hides the warning, and brings us down to 6 warnings for a make
defconfig bzImage.

BKrev: 3ec68b70RFPGEMyonowU_5sArHRmIg