[PATCH] fix oops in namespace.c
A familar type of Oops: d_path() can return an error ENAMETOOLONG, and
if we fail to test a segfault occurs.

So we must test.  What we do is a different matter.  Rather arbitrarily
I return the string " (too long)" for use in /proc/mounts.

BKrev: 3ec923339S46-hAGCoE0La2IWr3ETw