[PATCH] kNFSd: SVC sockets don't disable Nagle
I noticed that the Nagle algorithm was disabled just recently on the
client side, while it still seems to be enabled on the server side. The
previous patch came from Chuck Lever.

Below patch disables it on the server side as well. For latency reasons,
this should be the desired behaviour NFS at both client and server.

BKrev: 3ecaf0570Nz15ZZ51iLVdkUABxY-Yw