[PATCH] kNFSd: Make sure an RPC socket is closed immediately when a server write fails.
The check for "write was incomplete" is not atomic with the actual write,
so a second write could slip in after a partial write to a tcp socket
and this will cause problems.

So we move the locking out a bit, and test if the socket has
been deleted (SK_DEAD) between getting the lock and trying
to send.

BKrev: 3ecaf071PitTMR4Y5IJjpaeSU8KwYw