[PATCH] kNFSd: Allow request for nfsv4 pseudo root to perform an upcall.
Most cache-lookups to find export information will cause
an upcall, but currently the lookup to find the pseudo root
for nfsv4 wont asthe  cache-handle isn't made available.
This patch makes it available and makes sure that if the
upcall is made, the request is dropped (for now).

BKrev: 3ecaf091WwPq9-DsBxTVdongWG2yWg