[PATCH] Remove floating point use in cpia.c
From: Duncan Haldane <f.duncan.m.haldane@worldnet.att.net>

Gerd Knorr noticed a small use of floating point math in the cpia driver
updates for 2.5.x I sent you a while ago, and this is not allowed in the
kernel.

This was in some code taken essentially verbatim from the Windows cpia
driver released under the GPL by STM inc.  that had been incorporated in
the later versions of the cpia drivera sourceforge.

It turns out that the use of floating point was quite inessential, and I've
reimplemented the couple of lines of code involved in integer arithmetic.

BKrev: 3ed14025HwJ_M-rUWY8ccq8HHmCdZQ