Merge raven.il.steeleye.com:/home/jejb/BK/scsi-misc-2.5 into raven.il.steeleye.com:/home/jejb/BK/scsi-for-linus-2.5
2003/05/26 18:34:05-05:00 jejb
qla1280: convert the driver to the new SCSI error handler

2003/05/26 18:31:42-05:00 jejb
qla1280: set the data direction correctly

The data direction for the SCSI command was being strangely deduced
by checking for an old WRITE_6 command.

2003/05/26 18:28:45-05:00 jejb
qla1280: convert the driver to be endian neutral

This makes the driver compile and run on the PARISC platform.
The biggest issue for this driver is the firmware layout because
the chip insists on reading from firmware a 16 bit word at a time,
so the entire byte layout of the firmware had to be reversed.

2003/05/26 16:53:50-05:00 jejb
Merge ssh://linux-scsi@linux-scsi.bkbits.net/scsi-misc-2.5
into raven.il.steeleye.com:/home/jejb/BK/scsi-misc-2.5

2003/05/26 16:47:48-05:00 jejb
scsi sysfs add attribute release function.

2003/05/26 16:45:05-05:00 jejb
Fix up 53c700 compile

2003/05/26 12:36:59-05:00 James.Bottomley
[PATCH] Use of the new attribute modifiers on the 53c700

The attached patch illustrates how the new attribute functions can be
used.

The 53c700 modifies the queue_depth attribute so it can be set by the
root user, and also adds a new active_tags attribute which may be used
to obtain the information it currently puts out via its proc interface.

This achieves all I think we really need, which is the ability to have
LLDs provide their own routines for setting the default attributes.

2003/05/26 12:14:38-05:00 James.Bottomley
[PATCH] SCSI: Make sysfs attributes mutable

The attached patch makes the sysfs attributes a property of the host
template (I can't really see a reason why the actual attributes need to
change per host, so I think the template is the better place for them).

I've also provided helper functions to modify the attributes in the LLD
init routines.

2003/05/23 19:10:19-04:00 oliver
[PATCH] improve Documentation for DC395

this improves the Documentation for the DC395 driver for 2.5.
 dc395x.txt |   94 ++++++++++++++++++++++++++++++++++++++++---------------------
 1 files changed, 62 insertions(+), 32 deletions(-)

2003/05/23 11:14:10-04:00 Eddie.Williams
[PATCH] scsi: allow devices to restrict start on add

When a SCSI disk is added and it returns a NOT READY the SD driver is
automatically sending a START_UNIT command to spin the device up.  While this
may be the desired behavior for many if not most devices not all devices
either want or need this.  The attached patch provides a mechanism via the
device_list that allows a device to be defined to disable the automatic start
being issued on an add.

The patch also modifies the device_list for several devices that would prefer
to not have the start command issued.

2003/05/23 11:13:22-04:00 Andries.Brouwer
[PATCH] scsi: ten -> use_10_for_rw / use_10_for_ms

In the old days, ancient scsi devices understood 6-byte commands
and more recent ones also understood 10-byte commands.
Thus, we had a "ten" flag indicating that 10-byte commands worked.

These days, especially for usb-storage devices, the opposite
sometimes holds - 10-byte commands are supported, but 6-byte commands
are not.

The patch below changes the field ten into the pair of fields
use_10_for_rw, use_10_for_ms set initially when the driver
thinks these are supported. Ifthe device returns ILLEGAL_REQUEST
they are cleared.

This patch obsoletes a large amount of code in usb-storage,
and not only that, once the subsequent patch removes all this
usb-storage code many devices will work that hang today.


Andries

2003/05/23 11:12:45-04:00 Andries.Brouwer
[PATCH] isa_writeb args interchanged

2003/05/21 12:54:45-05:00 jejb
Fix while in spinup loop of sd

We must loop on UNIT ATTENTION sense (to clear the condition).  However, the code
is assuming that a sense condition will always be accompanied by a not good status.
We have several drivers in the tree that return GOOD with DRIVER_SENSE, so make the
loop condition reflect this.

2003/05/20 16:52:17-05:00 jejb
Fix up proc_info conversion in 53c700

The host variable is now passed in, so we shouldn't have a
similar variable defined in the proc_info routine

2003/05/20 14:52:01-05:00 davej
[PATCH] AM53C974 request region.

Request port before using it.

2003/05/19 21:41:14-05:00 jejb
Merge raven.il.steeleye.com:/home/jejb/BK/scsi-tmp-2.5
into raven.il.steeleye.com:/home/jejb/BK/scsi-misc-2.5

2003/05/19 21:38:46-05:00 hch
[PATCH] give ->proc_info a struct Scsi_Host * parameter

Yeah, I know the method is obsolete but for proper refcounting we
need to get rid of all those scsi_host_hn_get() abusers.

scsi_host_hn_get() and scsi_host_put() are not exported anymore
after this patch.

2003/05/19 21:17:03-05:00 dougg
[PATCH] scsi_mid_low_api.txt in lk 2.5.69

This is a catch up, applying the patch that Andries
Brouwer sent a while back. It will apply to lk 2.5.69
and 2.5.69-bk10 .

2003/05/19 21:08:03-05:00 oliver
[PATCH] fix irq handling for DC395

this fix makes sure that all test in irq handling in that
driver are for NULL thus fixing an oops.

2003/05/19 21:07:06-05:00 andmike
[PATCH] Call release on scsi legacy LLDD

In the sysfs scsi host reference counting merge the call to a legacy hosts
release function was removed.

This patch adds the call back in.

This patch was tested with insmod / rmmod on LLDDs using the legacy interface
and LLDDs using the new interface:

Legacy Interface:
qla2xxx-v8.00.00b1
qlogicisp
aic7xxx_old (Does not call scsi_unregister in release, but unloaded)

New Interface:
scsi_debug
aic7xxx (Many illegal context warnings caused by mid vs LLDD primitive
         mixing, but appears to unload clean).

 drivers/scsi/hosts.c |    4 ++++
 1 files changed, 4 insertions(+)

BKrev: 3ed2c57erYwVKzvClcCbpaZ9VyMr_A